Fix an issue with double "//" for absolute paths
Previously the path_prefix was treated like an ordinary file path part.
This lead to an additional '/' being interpolated between the "file" '/'
and another actual file in the root directory.

Now the path_prefix is handled separately, and prepended to the path
being constructed only at the top level call of walk.