Dash parser fixes (#18)
* Handle malformed substitutions gracefully

Instead of blowing up the stack, let the user know that there was an
error.

* Fix memory handling for nested evals.

Smoosh mishandled Dash's memory routines that could lead to buffer
overflows when executing nested evals.

After an eval is done, the stackmark must be popped before the file is
popped and the input string is unallocated. Skipping the stack popping
can lead to situations where Dash mistakenly believes that the free
space spans multiple stack blocks.

This change makes Smoosh pop the stackmark before unallocating.