Add escaped characters to the AST (#36)
To represent characters escaped by a backslash, a new AST entry is
added. This improves their handling considerably compared to the
previous state, where the escaping was in some cases ignored and in
other cases turned into a literal backslash.

The change has not yet been pushed "all the way through": `expanded_words` and `intermediate_fields` still use backslashes for escaping. That fix is a longer-term issue (#37).